Litmatch/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
sentence-transformers==2.2.2
numpy==1.26.4